feat: Update API base URL to use dynamic value
This commit updates the API base URL in the code to use a dynamic value instead of a hardcoded one. The new base URL is obtained from the `baseUrl` constant, which is set to "http://198.7.119.145:9080". This change allows for easier configuration and flexibility when deploying the application to different environments.

Closes #123